pyproject.toml
[tool.poetry]
name = "course_work5"
version = "0.1.0"
description = ""
authors = ["igorzinoviev08 <zinovievigor08@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"
pandas = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
